Logging has been refactored.
A new module has been added to handle logging. The main goal was to
provide more granular logging options, both for the logger itself and
for the formatting mechanism.

Note: This was a more difficult change than expected. It is likely
there are some residual issues to deal with. Specific logging changes
can be handled as various modules are refactored.