TDL Walker has become just a Walker.
The name change was useful to make sure it was clear that it was not the
TDL being "walked", but rather then AST. Since the logic is in the AST
folder, it seemed prudent to simply shorten the name. The methods used
also were changed to be a little more reflective of what they do.