Specs loader refactored.
One simple refactoring was changing this to a spec (singular) loader.
This more accurately reflects what the class is doing. The class itself
was taken out of the runtime folder although not out of the Runtime
class structure. This is in preparation for changing the concept of
the runtime to a context.