Runtime has become a context loader.
The previous class of Runtime has been refactored as a ContextLoader. This
is more in line with what Lucid is actually doing. Lucid is predicated upon
an operating context and a large part of its operation is loading that
context. Technically, this is a runtime, but it feels better to me to
describe this functionality less generically.