Reverse merge now avoids context paths.
A problem was happening where Lucid would not be able to run certain
repositories if a profile was used. This problem would not occur if
profiles were not used. The root cause of the issue was that the
reverse merge portion of the options handling was a little too
aggressive in assumptions made about the context paths, such as
definitions path, steps path, etc. That has now been fixed. There
is a possibility that this will impact how those elements can be
specified at the command line.