Unit testing support added.
RSpec support has been added for unit testing the Lucid framework. This
ties into code coverage and reporting tools. There is a potential issue
here if I have an acceptance test example as part of the code repository.
The issue is that Lucid will be hooking into RSpec and that may conflict
with the unit test execution.