Lucid is able to parse Feature elements.
The RSpec usurping process calls upon a method that seeks out test spec
files and then uses the Gherkin API to build an AST representation of the
test spec. This representation will be stored in objects that can be
referenced and used in RSpec example groups. The current commit is able
to recognize a high-level test spec element (Feature or Ability).