Lucid is able to parse Background elements.
This commit recognizes background elements within a test spec. Backgrounds
are basically just another scenario from a parsing perspective. They have
an identifier, an optional name, as well as a set of steps. As with the
scenario element, all steps for a background are stored in a context.
Backgrounds are added to the feature currently being processed just as
scenarios are.