Lucid is able to parse Scenario Outline elements.
This commit adds support for scenario outlines (also known as scenario
templates). A scenario outline is a parameterized scenario that utilizes
a data table for its conditions. This means a scenario outline has to be
parsed as a specific element that can hold a collection of steps. These
steps will have placeholders. An outline will also have a collection of
examples, which are essentially scenarios. These scenarios provide specific
data that will replace the placeholders. This means each example row must
be turned into a scenario element.

Supporting this change required adding support for steps in non-outline
scenarios that hold data tables. Lucid provides a specific Table class
to make operations with and on tables easier.