Lucid is able to parse data tables that use an example table.
This commit adds support for a variation on using scenario outlines. It is
possible to have a step with a data table where the data table itself has
rows that refer to the example table of a scenario outline. In these cases
it is necessary to have the data table populated with values from the
examples table during each run of a scenario execution.