Lucid is able to parse DocString elements.
This commit allows Lucid to recognize a type of step argument known as a
doc string. The handling for this is very similar to handling data tables,
in that both are arguments to a step. This commit also makes sure that
scenario outlines which have their parameter in a doc string can have the
correct value from the examples table substituted into the doc string's
parameter.