Structure for running individual step from spec.
This commit puts in the basis for how Lucid will operate. The idea is that
the SpecRunner is loaded into the RSpec configuration. After specs have
been loaded (via the SpecLoader) a run() method is called for any example
block, which in RSpec corresponds to an it() block. That run() method is in
the SpecRunner(). The SpecRunner mixes in the StepRunner. That StepRunner
has a step() method that handles actually executing a given step that is
read in from a scenario in the test spec.

Currently this structure is just a shell implementation.