Lucid treats non-matched steps as pending.
In order for a step to execute, it must have a matcher. A matcher is what
ties the test spec step to actual code that will execute when that step is
processed. If such a matcher is not found, that means the step is pending
and must be skipped. However, in order to do that it's necessary to search
for a matcher by some common mechanism. That mechanism will be to look for
methods that have the prefix "matcher: ".